Refine extension SDK internals
