+++ Packages/CmuxSidebarProviderKit/Sources/CmuxSidebarProviderKit/CmuxSidebarProviderPresentation.swift
@@ -1,0 +1,12 @@
+import Foundation
+
+public enum CmuxSidebarProviderPresentation: String, Codable, Equatable, Sendable {
+    case tree
+    case browserStack = "browser-stack"
+}
+
+public enum CmuxSidebarProviderWorkspacePopoverTab: String, Codable, CaseIterable, Equatable, Sendable {
+    case notes
+    case browser
+    case pullRequest
+}
